upload-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
minio==7.1.0
python-dotenv==0.19.0
cassandra-driver==3.28.0
PyJWT==2.8.0
flask==2.3.3
flask-cors==4.0.0
